PuzzleList with hashing and hScore storage
Added puzzleList class that stores main data about puzzle states, so the hash value and the hScore (result of heuristic), so there is no need to count hScore on each iteration. It sure does not have a big impact on speed, but it is better to some extend. Also increasing bound by 10, to just try it out and see how it influences the speed. Same, still searching for ways to solve cases when the time is too long