Autosolve-Click&Move are synchronized
So it is a but complicated, but I am using clickedStated as a semafore a little, so it has three main values 1 for primary clicked state, 2 for changed-after-clicked-and-processed state, and False for just passive button. Also separated button into its own file. Astar is deleted because puzzleStar has all the functionality from it, but is enclosed in a class