ShuffleMoves will shuffle from start
So I have noticed IDA star neing way slower on the second, third and so on shuffle. Way slower. So I was shuffling an already shuffled board, thus the complexity grew. It is real good tho that my code even with that seemed to work, but I lost the shuffle control, that really has the impact on speed. The puzzle shuffled with 200 shuffles may be as complecated as the one with 1000 shuffles, but still it is not the point to make the starting configuration the most complicated game ever.