Save game state to file
Now the game starting screen in a non-active game state. It includes button for resuming the saved game, which uses readBoard function from File Module (which will later also be a class). In the active state there is a Save game button, for yep saving the game. However, when user exits, the state of the game is also saved (thinking if it will not make conflict with user expectations.) Added backToInit method to button, which just returns it to its creation state.