New style, better PuzzleList dict usage
So the colors have changed. I reset score after shuffle and autosolving. In PuzzleStar in IDA star and search we no longer store nodes in a list, but insted we immediately add them to the puzzle_list, which stores the hScore and hash value as a key in ordered dictionary. This seems to speed things up, because time for is in operation in list is O(n), which is too much. Also delegetade the direction count to the moveTiles method